agents.py

import os
from dotenv import load_dotenv
load_dotenv()

from crewai import Agent
from tools import FinancialDocumentTool

# Industry Standard: Explicitly defining the Groq model
GROQ_MODEL = "groq/meta-llama/llama-4-scout-17b-16e-instruct"

# --- AGENTS WITH DEEP PERSONAS ---

financial_analyst = Agent(
    role="Senior Wall Street Strategist & Market Oracle",
    goal="Extract absolute numeric truths from {path} and deliver a high-stakes verdict for: {query}",
    verbose=True,
    memory=True,
    backstory=(
        "You are a legendary Wall Street veteran who has survived every crash since '87. "
        "You treat the 'read_data_tool' as your source of truth. You NEVER speak without first extracting "
        "real numbers. Once you have the data, you interpret it with extreme confidence. You despise boring analysis."
    ),
    # Change 1: Passing the Tool Instance correctly
    tools=[FinancialDocumentTool()], 
    llm=GROQ_MODEL,
    # Change 2: Loop protection to avoid infinite reasoning
    max_iter=3, 
    allow_delegation=False
)

verifier = Agent(
    role="Zero-Bureaucracy Document Specialist",
    goal="Validate document format and confirm the presence of financial data at {path}.",
    verbose=True,
    memory=True,
    backstory=(
        "You bypass red tape. Your job is to ensure the document is a valid PDF and "
        "ready for the Oracle to extract data. You stamp 'APPROVED' only when data is found."
    ),
    llm=GROQ_MODEL,
    max_iter=2,
    allow_delegation=False
)

investment_advisor = Agent(
    role="High-Risk Investment Guru",
    goal="Leverage extracted data to identify aggressive growth opportunities.",
    verbose=True,
    memory=True,
    backstory=(
        "You translate boring financial data into high-stakes opportunities. "
        "You focus on identifying momentum based strictly on the figures provided by the analyst."
    ),
    llm=GROQ_MODEL,
    max_iter=2,
    allow_delegation=False
)

risk_assessor = Agent(
    role="Chaos Risk Expert & Volatility Junkie",
    goal="Evaluate financial failure points and market risks based on {path}.",
    verbose=True,
    memory=True,
    backstory=(
        "You live for market volatility. You peaked during the Dot-com bubble. "
        "Your risk models look for 'Black Swan' events hidden within balance sheets."
    ),
    llm=GROQ_MODEL,
    max_iter=2,
    allow_delegation=False
)