requirements.txt
# click==8.1.7
# # Keep crewai at version 0.130.0
# # Other package versions are flexible and can be changed
# # Only change crewai version if there are critical dependency conflicts that cannot be resolved by other means
# crewai==0.130.0 
# crewai-tools==0.47.1
# fastapi==0.110.3
# google-ai-generativelanguage==0.6.4
# google-api-core==2.10.0
# google-api-python-client==2.131.0
# google-auth==2.29.0
# google-auth-httplib2==0.2.0
# google-cloud-aiplatform==1.53.0
# google-cloud-bigquery==3.23.1
# google-cloud-core==2.4.1
# google-cloud-resource-manager==1.12.3
# google-cloud-storage==2.16.0
# google-crc32c==1.5.0
# google-generativeai==0.5.4
# google-resumable-media==2.7.0
# googleapis-common-protos==1.63.0
# Jinja2==3.1.4
# jsonschema==4.22.0
# langchain-core==0.1.52
# langsmith==0.1.67
# numpy==1.26.4
# oauthlib==3.2.2
# onnxruntime==1.18.0
# openai==1.30.5
# opentelemetry-api==1.25.0
# opentelemetry-exporter-otlp-proto-common==1.25.0
# opentelemetry-exporter-otlp-proto-grpc==1.25.0
# opentelemetry-exporter-otlp-proto-http==1.25.0
# opentelemetry-instrumentation==0.46b0
# opentelemetry-instrumentation-asgi==0.46b0
# opentelemetry-instrumentation-fastapi==0.46b0
# opentelemetry-proto==1.25.0
# opentelemetry-sdk==1.25.0
# opentelemetry-semantic-conventions==0.46b0
# opentelemetry-util-http==0.46b0
# pandas==2.2.2
# pillow==10.3.0
# pip==24.0
# protobuf==4.25.3
# pydantic==1.10.13
# pydantic_core==2.8.0

# # Missing Dependencies added for PDF processing and Gemini Integration
# PyPDF2==3.0.1
# langchain-google-genai==1.0.3
# python-dotenv==1.0.1
# uvicorn==0.29.0
# python-multipart==0.0.9



# ═══════════════════════════════════════════════════════════════
# requirements.txt  —  Enterprise Financial Engine
# ───────────────────────────────────────────────────────────────
# Files: main.py · agents.py · task.py · tools.py
#        database.py · streamlit_app.py
#
# Install: pip install -r requirements.txt
# ═══════════════════════════════════════════════════════════════


# ── CrewAI — PINNED, DO NOT UPGRADE ──────────────────────────
# crewai 0.130.0 requires pydantic v1 (NOT v2)
# Upgrading crewai will break Agent, Task, BaseTool, Crew APIs
crewai==0.130.0
crewai-tools==0.47.1


# ── Pydantic — PINNED v1 (crewai 0.130.0 hard requirement) ───
# ⚠️  DO NOT upgrade to pydantic>=2.0.0 — crewai will break
pydantic==1.10.13
pydantic_core==2.8.0


# ── FastAPI Backend (main.py) ─────────────────────────────────
# BackgroundTasks, CORSMiddleware, File, UploadFile, Form
fastapi==0.110.3
uvicorn==0.29.0
python-multipart==0.0.9          # Required for UploadFile / PDF upload


# ── Streamlit Frontend (streamlit_app.py) ─────────────────────
streamlit>=1.32.0
requests>=2.31.0


# ── Groq LLM (agents.py) ─────────────────────────────────────
# GROQ_MODEL = "groq/meta-llama/llama-4-scout-17b-16e-instruct"
# crewai routes groq/ prefix calls via litellm
groq>=0.9.0
litellm>=1.40.0


# ── PDF Processing (tools.py — FinancialDocumentTool) ─────────
PyPDF2==3.0.1


# ── Environment (.env → GROQ_API_KEY) ────────────────────────
python-dotenv==1.0.1


# ── LangChain (crewai 0.130.0 internal dependency) ───────────
langchain-core==0.1.52
langchain>=0.1.20
langsmith>=0.1.67


# ── OpenAI SDK (crewai uses internally) ──────────────────────
openai==1.30.5


# ── OpenTelemetry (crewai 0.130.0 requires all of these) ─────
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0


# ── Data & Utilities ─────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
Jinja2==3.1.4
click==8.1.7
jsonschema==4.22.0
pillow==10.3.0
protobuf==4.25.3


# ── HTTP & Async ─────────────────────────────────────────────
httpx>=0.27.0
anyio>=4.3.0


# ── Google Auth (crewai + googleapis indirect deps) ───────────
google-auth==2.29.0
googleapis-common-protos==1.63.0
google-api-core==2.10.0


# ── ONNX Runtime (crewai embeddings / tokenizer) ─────────────
onnxruntime==1.18.0


# ── sqlite3 ───────────────────────────────────────────────────
# database.py → sqlite3 is Python built-in — NO pip install needed
# financial_data.db auto-creates on first POST /analyze